backend/requirements.txt
asgiref==3.10.0
channels==4.0.0
channels-redis==4.1.0
daphne==4.0.0
diff-match-patch==20241021
Django==4.2.25
django-allauth
django-cors-headers==4.9.0
django-import-export==4.3.12
django-redis==6.0.0
djangorestframework==3.16.1
djangorestframework-simplejwt
dj-rest-auth
pillow==11.3.0
psycopg2-binary==2.9.11
PyJWT
python-dotenv
qrcode==7.4.2
redis==5.0.1
shortuuid==1.0.11
sqlparse==0.5.3
tablib==3.9.0
typing_extensions==4.15.0
